Consolidate hash functions into their own file
