+++ src/core/login/recovery2.js
@@ -2,7 +2,8 @@
 
 import { base64 } from 'rfc4648'
 
-import { decrypt, encrypt, hmacSha256 } from '../../util/crypto/crypto.js'
+import { decrypt, encrypt } from '../../util/crypto/crypto.js'
+import { hmacSha256 } from '../../util/crypto/hashes.js'
 import { fixOtpKey, totp } from '../../util/crypto/hotp.js'
 import { utf8 } from '../../util/encoding.js'
 import { type ApiInput } from '../root-pixie.js'
